RSM to handle long valid topic names (#5)
The previous implementation did not handle long topic names because the topic name was encoded (potentially lossy) in the container name. Container names are restrictive and only 63 chars long. This mapping does not use topic name in the container name. It uses an optional prefix and the canonical representation of the topic UUID. The overall mapping is as follows. It's acknowledged that it's not very operationally friendly because the topic name is nowhere to be found.

|----------- container name ------------|----------------blob name---------------------|
|OptionalContainerPrefix-TopicIdInBase36|PartitionNum.LogSegmentCanonicalUUID.IndexType|