Deleting One Segment Does Not Delete Other Segments (#10)
A previous PR allowed creating remote log segments for topics with long names. This was achieved by using a single container for all the topic-partitions of a given topic. The previous mapping used a single container for a given topic-partition. The previous PR changed the mapping but did not change the deletion logic. The deletion logic deleted the entire container, which deleted not only all the old segments but deleted everything (effectively delete topic). That's wrong.

This PR adds the following.
# Resolves the above issue by deleting the segments and index files one-by-one. It's also essential for Kafka's topic retention logic.
# The transaction index deletion checks if it exists because in Linkedin we don't use transactions.
# Added a new unit test verifies that deleting older segment does not delete anything more than necessary.
# Added informative logging while handling the exception from the Azure API.
# Using simplified blob client instead of blockblob client as deletion is a common operation.